pyproject.toml
[tool.poetry]
name = "dip-coater"
version = "1.1.0"
description = "A terminal / Textual app to control a dip coater motor"
authors = [
    "Rik Huygen <rik.huygen@kuleuven.be>",
    "Sibo Van Gool <sibo.vangool@kuleuven.be>"
]
license = "MIT"
readme = "README.md"
packages = [{include = "dip_coater", from = "src"}]
package-mode = true

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
textual = ">=0.73.0"
textual-dev = "*"
uvloop = ">=0.19.0"
pytrinamic = "^0.2.11"
tmc-2209-raspberry-pi = ">=0.5.2"
mock-gpio = "^0.1.8"
flake8 = "^7.1.0"

[tool.poetry.group.rpi]
optional = true

[tool.poetry.group.rpi.dependencies]
"RPi.GPIO" = ">=0.7.1"

[tool.poetry.scripts]
dip-coater = "dip_coater.app:main"

[tool.poetry.urls]
repository = "https://github.com/IvS-KULeuven/dip_coater"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"